feat: 구입 금액 검증 로직 PurchaseInformation 으로 이관
